+++ priv/repo/migrations/20190222114737_add_dynamic_form_fields.exs
@@ -1,0 +1,10 @@
+defmodule TdDq.Repo.Migrations.AddDynamicFormFields do
+  use Ecto.Migration
+
+  def change do
+    alter table("rules") do
+      add :df_name, :string
+      add :df_content, :map
+    end
+  end
+end